instructions, bug fixes, some other things too that i forgot about 	modified:   src/main/java/com/bitdecay/game/component/DamageComponent.java 	modified:   src/main/java/com/bitdecay/game/component/HealthComponent.java 	modified:   src/main/java/com/bitdecay/game/gameobject/GameObjectFactory.java 	modified:   src/main/java/com/bitdecay/game/room/DemoRoom.java 	new file:   src/main/java/com/bitdecay/game/screen/InstructionsScreen.java 	modified:   src/main/java/com/bitdecay/game/screen/MainMenuScreen.java 	modified:   src/main/java/com/bitdecay/game/system/FuelGaugeSystem.java 	modified:   src/main/java/com/bitdecay/game/system/HealthSystem.java 	modified:   src/main/java/com/bitdecay/game/system/InvincibleComponent.java 	modified:   src/main/java/com/bitdecay/game/system/InvincibleSystem.java 	modified:   src/main/java/com/bitdecay/game/ui/Fuel.java 	modified:   src/main/resources/conf/application.conf 	new file:   src/main/resources/conf/instructions.conf

+++ src/main/java/com/bitdecay/game/component/DamageComponent.java
@@ -5,9 +5,9 @@ package com.bitdecay.game.component;
  */
 public class DamageComponent extends AbstractComponent {
 
-    public int damage;
+    public float damage;
 
-    public DamageComponent(int dmg){
+    public DamageComponent(float dmg){
         damage = dmg;
     }
 }
